Move util funcs
